Added dev mode
